#17 #18 #49 created BaseAgentCondition, ClosestCondition, CollisionCondition, RangeCondition, and changed IntervalCondition to extend Condition rather than implement it since it was changed to an abstract class (and changed Agent's direction to be a double)

+++ gameengine/src/state/agent/Agent.java
@@ -23,7 +23,7 @@ public class Agent implements IAgent, Cloneable {
     private int health;
     private int width;
     private int height;
-    private int direction;
+    private double direction;
 
     protected List<ActionDecision> actionDecisions;
